+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/person/ai/mission/meta/CollectRegolithMeta.java
@@ -13,6 +13,7 @@ import org.mars_sim.msp.core.person.ai.job.JobType;
 import org.mars_sim.msp.core.person.ai.mission.CollectRegolith;
 import org.mars_sim.msp.core.person.ai.mission.Mission;
 import org.mars_sim.msp.core.person.ai.mission.MissionType;
+import org.mars_sim.msp.core.person.ai.mission.MissionUtil;
 import org.mars_sim.msp.core.person.ai.mission.VehicleMission;
 import org.mars_sim.msp.core.person.ai.role.RoleType;
 import org.mars_sim.msp.core.structure.Settlement;
@@ -57,7 +58,7 @@ public class CollectRegolithMeta extends AbstractMetaMission {
 					|| RoleType.SUB_COMMANDER == roleType
 					) {
 
-	    		int numEmbarked = VehicleMission.numEmbarkingMissions(settlement);
+	    		int numEmbarked = MissionUtil.numEmbarkingMissions(settlement);
 	    		int numThisMission = missionManager.numParticularMissions(MissionType.COLLECT_REGOLITH, settlement);
 
 		   		// Check for # of embarking missions.
