+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/person/ai/mission/meta/ExplorationMeta.java
@@ -16,6 +16,7 @@ import org.mars_sim.msp.core.person.ai.job.JobType;
 import org.mars_sim.msp.core.person.ai.mission.Exploration;
 import org.mars_sim.msp.core.person.ai.mission.Mission;
 import org.mars_sim.msp.core.person.ai.mission.MissionType;
+import org.mars_sim.msp.core.person.ai.mission.MissionUtil;
 import org.mars_sim.msp.core.person.ai.mission.RoverMission;
 import org.mars_sim.msp.core.person.ai.mission.VehicleMission;
 import org.mars_sim.msp.core.person.ai.role.RoleType;
@@ -73,7 +74,7 @@ public class ExplorationMeta extends AbstractMetaMission {
 				}
 
 				missionProbability = 1D;
-				int numEmbarked = VehicleMission.numEmbarkingMissions(settlement);
+				int numEmbarked = MissionUtil.numEmbarkingMissions(settlement);
 				int numThisMission = missionManager.numParticularMissions(MissionType.EXPLORATION, settlement);
 
 		   		// Check for # of embarking missions.
