+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/person/ai/mission/meta/FieldStudyMeta.java
@@ -15,6 +15,7 @@ import org.mars_sim.msp.core.person.Person;
 import org.mars_sim.msp.core.person.ai.job.JobType;
 import org.mars_sim.msp.core.person.ai.mission.FieldStudyMission;
 import org.mars_sim.msp.core.person.ai.mission.MissionType;
+import org.mars_sim.msp.core.person.ai.mission.MissionUtil;
 import org.mars_sim.msp.core.person.ai.mission.RoverMission;
 import org.mars_sim.msp.core.person.ai.mission.VehicleMission;
 import org.mars_sim.msp.core.person.ai.role.RoleType;
@@ -64,7 +65,7 @@ public class FieldStudyMeta extends AbstractMetaMission {
 
 				missionProbability = 1D;
 				
-				int numEmbarked = VehicleMission.numEmbarkingMissions(settlement);
+				int numEmbarked = MissionUtil.numEmbarkingMissions(settlement);
 				int numThisMission = missionManager.numParticularMissions(mType, settlement);
 				
 		   		// Check for # of embarking missions.
