+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/person/ai/mission/meta/MiningMeta.java
@@ -17,6 +17,7 @@ import org.mars_sim.msp.core.person.ai.job.JobType;
 import org.mars_sim.msp.core.person.ai.mission.Mining;
 import org.mars_sim.msp.core.person.ai.mission.Mission;
 import org.mars_sim.msp.core.person.ai.mission.MissionType;
+import org.mars_sim.msp.core.person.ai.mission.MissionUtil;
 import org.mars_sim.msp.core.person.ai.mission.RoverMission;
 import org.mars_sim.msp.core.person.ai.mission.VehicleMission;
 import org.mars_sim.msp.core.person.ai.role.RoleType;
@@ -75,7 +76,7 @@ public class MiningMeta extends AbstractMetaMission {
 	            if (!Mining.areAvailableAttachmentParts(settlement))
 	            	return 0;
 
-				int numEmbarked = VehicleMission.numEmbarkingMissions(settlement);
+				int numEmbarked = MissionUtil.numEmbarkingMissions(settlement);
 				int numThisMission = missionManager.numParticularMissions(MissionType.MINING, settlement);
 
 		   		// Check for # of embarking missions.
