+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/person/ai/task/UnloadVehicleGarage.java
@@ -291,23 +291,21 @@ public class UnloadVehicleGarage extends Task implements Serializable {
 				VehicleMission vehicleMission = (VehicleMission) mission;
 
 				if (vehicleMission.isVehicleUnloadableHere(settlement)) {
-					if (vehicleMission.hasVehicle()) {
-						Vehicle vehicle = vehicleMission.getVehicle();
-						if (vehicle instanceof Crewable) {
-							Crewable c = (Crewable)vehicle;
-							if (c.getCrewNum() > 0 || c.getRobotCrewNum() > 0) {
-								// Has occupants so skip it
-								continue;
-							}
+					Vehicle vehicle = vehicleMission.getVehicle();
+					if (vehicle instanceof Crewable) {
+						Crewable c = (Crewable)vehicle;
+						if (c.getCrewNum() > 0 || c.getRobotCrewNum() > 0) {
+							// Has occupants so skip it
+							continue;
 						}
+					}
 
-						// If looking for garaged vehicles; then add to garage otherwise
-						// check vehicle is not in garage
-						if (!isFullyUnloaded(vehicle)) {
-							if ((addToGarage && settlement.getBuildingManager().addToGarage(vehicle))
-									|| (!addToGarage && !settlement.getBuildingManager().isInGarage(vehicle))) {
-								result.add(vehicleMission);
-							}
+					// If looking for garaged vehicles; then add to garage otherwise
+					// check vehicle is not in garage
+					if (!isFullyUnloaded(vehicle)) {
+						if ((addToGarage && settlement.getBuildingManager().addToGarage(vehicle))
+								|| (!addToGarage && !settlement.getBuildingManager().isInGarage(vehicle))) {
+							result.add(vehicleMission);
 						}
 					}
 				}
