+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/vehicle/VehicleAirlock.java
@@ -253,7 +253,7 @@ extends Airlock {
 	 * @return
 	 */
 	public Person getAssociatedPerson(int id) {
-		return vehicle.getAssociatedSettlement().getAssociatedPerson(id);
+		return unitManager.getPersonByID(id);
 	}
 
 	/**
