+++ mars-sim-ui/src/main/java/org/mars_sim/msp/ui/swing/tool/mission/edit/InfoPanel.java
@@ -32,9 +32,11 @@ import org.mars_sim.msp.core.Coordinates;
 import org.mars_sim.msp.core.Simulation;
 import org.mars_sim.msp.core.UnitManager;
 import org.mars_sim.msp.core.person.Person;
+import org.mars_sim.msp.core.person.ai.mission.AbstractVehicleMission;
 import org.mars_sim.msp.core.person.ai.mission.CollectResourcesMission;
 import org.mars_sim.msp.core.person.ai.mission.Mission;
 import org.mars_sim.msp.core.person.ai.mission.MissionPhase;
+import org.mars_sim.msp.core.person.ai.mission.MissionUtil;
 import org.mars_sim.msp.core.person.ai.mission.RoverMission;
 import org.mars_sim.msp.core.person.ai.mission.VehicleMission;
 import org.mars_sim.msp.core.person.ai.task.utils.Worker;
@@ -250,15 +252,15 @@ public class InfoPanel extends JPanel {
 			// Check if go home action can be added.
 			int nextNavpointIndex = vehicleMission.getNextNavpointIndex();
 			if ((nextNavpointIndex > -1) && (nextNavpointIndex < (vehicleMission.getNumberOfNavpoints() - 1))) {
-				if (!mission.getPhase().equals(VehicleMission.EMBARKING))
+				if (!mission.getPhase().equals(AbstractVehicleMission.EMBARKING))
 					actions.add(ACTION_HOME);
 			}
 
 			// Check if nearest settlement action can be added.
 			try {
-				Settlement closestSettlement = vehicleMission.findClosestSettlement();
+				Settlement closestSettlement = MissionUtil.findClosestSettlement(mission.getCurrentMissionLocation());
 				if ((closestSettlement != null) && !closestSettlement.equals(vehicleMission.getAssociatedSettlement())) {
-					if (!mission.getPhase().equals(VehicleMission.EMBARKING))
+					if (!mission.getPhase().equals(AbstractVehicleMission.EMBARKING))
 						actions.add(ACTION_NEAREST);
 				}
 			}
